Add funding model updates to support BIGINT columns for monetary values. Enhanced API and frontend to include 'is_guest' attribute in pledges, allowing differentiation between donations and pledges. Introduced My Pledges screen for users to view their contributions, with filtering and search capabilities. Updated event detail and pledge receipt screens to reflect changes in terminology and display logic for guest donations.

+++ Backend/app/models/funding.py
@@ -3,7 +3,7 @@ Funding / pledge model.
 """
 import enum
 from datetime import datetime
-from sqlalchemy import Boolean, Integer, String, DateTime, ForeignKey, Enum
+from sqlalchemy import BigInteger, Boolean, Integer, String, DateTime, ForeignKey, Enum
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -21,9 +21,9 @@ class Funding(Base):
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     event_id: Mapped[int] = mapped_column(ForeignKey("events.id"), nullable=False, index=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"), nullable=False, index=True)
-    amount_cents: Mapped[int] = mapped_column(Integer, nullable=False)
-    platform_cut_cents: Mapped[int] = mapped_column(Integer, nullable=False, default=0)  # platform commission on pledge
-    net_to_organizer_cents: Mapped[int] = mapped_column(Integer, nullable=False, default=0)  # amount - platform_cut
+    amount_cents: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    platform_cut_cents: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
+    net_to_organizer_cents: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
     status: Mapped[FundingStatus] = mapped_column(Enum(FundingStatus), nullable=False, default=FundingStatus.pledged)
     is_guest: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)  # guest pledges are non-refundable
     reserved_spots: Mapped[int] = mapped_column(Integer, nullable=False, default=0)  # spots reserved for future ticket purchase
